+++ src/lib.rs
@@ -201,7 +201,7 @@ pub mod auth;
 pub mod client;
 pub mod database;
 pub mod error;
-pub mod kv;
+pub mod kv1;
 pub mod kv2;
 pub mod pki;
 pub mod ssh;
